requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
validators==0.18.2
requests==2.28.1
SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
gevent==22.10.2

# flask-jsontools==0.1.7